+++ CHANGELOG.md
@@ -2,6 +2,60 @@
 
 ## [Unreleased]
 
+## [0.5.5]
+
+* Add `RiQueryStr` and `RiQueryString` types for query.
+* Add functions with `try_` prefix are introduced for normalization and
+  IRI resolution, and deprecate non-`try` versions.
+* Add encoder types for percent-encoding in `percent_encoding` module.
+
+### Added
+* Add `RiQueryStr` and `RiQueryString` types for query.
+* Add functions with `try_` prefix are introduced for normalization and resolution.
+    + List of added functions:
+        - `types::RiStr::try_normalize()`
+        - `types::RiStr::try_normalize_whatwg()`
+        - `types::RiAbsoluteStr::try_normalize()`
+        - `types::RiAbsoluteStr::try_normalize_whatwg()`
+        - `resolve::try_resolve()`
+        - `resolve::try_resolve_whatwg()`
+        - `resolve::try_resolve_normalize()`
+        - `resolve::try_resolve_normalize_whatwg()`
+        - `types::RiReferenceStr::try_resolve_against()`
+        - `types::RiReferenceStr::try_resolve_normalize_against()`
+        - `types::RiReferenceStr::try_resolve_whatwg_against()`
+        - `types::RiReferenceStr::try_resolve_normalize_whatwg_against()`
+        - `types::RiRelativeStr::try_resolve_against()`
+        - `types::RiRelativeStr::try_resolve_normalize_against()`
+        - `types::RiRelativeStr::try_resolve_whatwg_against()`
+        - `types::RiRelativeStr::try_resolve_normalize_whatwg_against()`
+* Add encoder types for percent-encoding in `percent_encoding` module.
+    + List of added types:
+        - `percent_encoding::PercentEncoded`
+        - `percent_encoding::PercentEncodedForIri` (type alias)
+        - `percent_encoding::PercentEncodedForUri` (type alias)
+
+### Changed (non-breaking)
+* Deprecate non-`try` function names for normalization and resolution.
+    + List of deprecated functions:
+        - `types::RiStr::normalize()`
+        - `types::RiStr::normalize_whatwg()`
+        - `types::RiAbsoluteStr::normalize()`
+        - `types::RiAbsoluteStr::normalize_whatwg()`
+        - `resolve::resolve()`
+        - `resolve::resolve_whatwg()`
+        - `resolve::resolve_normalize()`
+        - `resolve::resolve_normalize_whatwg()`
+        - `types::RiReferenceStr::resolve_against()`
+        - `types::RiReferenceStr::resolve_normalize_against()`
+        - `types::RiReferenceStr::resolve_whatwg_against()`
+        - `types::RiReferenceStr::resolve_normalize_whatwg_against()`
+        - `types::RiRiAblosuteStrStr::resolve_against()`
+        - `types::RiRiAblosuteStrStr::resolve_normalize_against()`
+        - `types::RiRiAblosuteStrStr::resolve_whatwg_against()`
+        - `types::RiRiAblosuteStrStr::resolve_normalize_whatwg_against()`
+    + Use functions with `try_` prefix instead.
+
 ## [0.5.4]
 
 * Implement IRI resolution and normalization that uses serialization algorithm
@@ -529,7 +583,8 @@ Beleive rustdoc rather than this CHANGELOG.**
 
 Totally rewritten.
 
-[Unreleased]: <https://github.com/lo48576/iri-string/compare/v0.5.4...develop>
+[Unreleased]: <https://github.com/lo48576/iri-string/compare/v0.5.5...develop>
+[0.5.4]: <https://github.com/lo48576/iri-string/releases/tag/v0.5.5>
 [0.5.4]: <https://github.com/lo48576/iri-string/releases/tag/v0.5.4>
 [0.5.3]: <https://github.com/lo48576/iri-string/releases/tag/v0.5.3>
 [0.5.2]: <https://github.com/lo48576/iri-string/releases/tag/v0.5.2>